requirements.txt
# Batch ingestion + bronze lakehouse (imports used under app/)
# Keep <2.2 so this file also works inside the apache/airflow:2.9.3 image.
pandas>=2.1.2,<2.2
requests>=2.31,<3
python-dotenv>=1.0,<2
python-json-logger>=2.0,<4

# --- Streaming (Binance WS -> Kafka) ---
kafka-python>=2.0,<3
websocket-client>=1.7,<2
yfinance>=0.2.40,<1

# --- Airflow ---
# apache-airflow>=2.10,<3

# --- Parquet ---
pyarrow>=15,<20

# --- Dashboard API (FastAPI) ---
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
pydantic>=2.9,<3

# --- Sentiment (FinBERT) ---
# Free, open-source, runs locally, no API key needed.
# Model (~440 MB) downloads once to ~/.cache/huggingface/ on first run.
transformers>=4.40,<5
torch>=2.2,<3